restructed working demo
